Migrates fixtures into conftest
