--- tests/git_actions/test_git_actions.py
@@ -1,13 +1,0 @@
-from pathlib import Path
-from time import sleep
-from unittest.mock import patch
-
-import git
-
-from . import build_dir
-from od_compiler.util.git_actions import updateOD
-from od_compiler.util.git_actions import updateSubmodules
-
-
-def test_update_od():
-    pass
