+++ tests/utilities/test_build.py
@@ -2,16 +2,13 @@ from pathlib import Path
 
 import pytest
 
-from . import make_run_dirs
-from . import runs_list
-from . import test_run_dir
 from od_compiler.util.utilities import cleanOldRuns
 from od_compiler.util.utilities import stageBuild
-from tests import build_dir
+from tests.utilities import runs_list
 
 
 def test_build_staging_create_files_without_main(build_dir: Path):
-    code = "world.log << 'Hello!'"
+    code = 'world.log << "Hello!"'
     expected_files = ["server_config.toml", "code.dm", "test.dme", "map.dmm"]
     stage_dir = build_dir.joinpath("staging")
     stage_dir.mkdir()
@@ -24,7 +21,7 @@ def test_build_staging_create_files_without_main(build_dir: Path):
 def test_build_staging_create_files_with_defined_main(build_dir: Path):
     code = """\
 /proc/example()
-  world.log << 'Hello!'
+  world.log << "Hello!"
 
 /proc/main()
   example()
@@ -44,7 +41,7 @@ def test_build_staging_verify_file_contents(build_dir: Path):
 /var/TEST/aaaaaa=new
 
 /proc/example()
-  world.log << 'Hello!'
+  world.log << "Hello!"
 
 /proc/main()
   example()
