+++ tests/utilities/test_logs.py
@@ -2,7 +2,6 @@ from pathlib import Path
 
 import pytest
 
-from . import build_dir
 from od_compiler.util.utilities import splitLogs
 from od_compiler.util.utilities import writeOutput
